Move out of models directory
